[MOD] : response transferRecent dto 필드 타입 변경

+++ app/src/main/java/com/sopt/kakaobank/data/dto/response/ResponseTransferRecentDto.kt
@@ -9,7 +9,7 @@ data class ResponseTransferRecentDto(
     @SerialName("accountName")
     val accountName: String,
     @SerialName("accountNumber")
-    val accountNumber: Int,
+    val accountNumber: Long,
     @SerialName("isAccountLike")
     val isAccountLike: Boolean,
     @SerialName("createdAt")
